@@ -6,6 +6,7 @@
 	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.1" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.Certificate" Version="6.0.8" />
 		<PackageReference Include="Microsoft.AspNetCore.Cors" Version="2.2.0" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.10" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.10">
@@ -19,6 +20,6 @@
 	</ItemGroup>
 	<PropertyGroup>
 		<!-- <TargetFramework>netcoreapp3.1</TargetFramework> -->
-		<TargetFramework>net5.0</TargetFramework>
+		<TargetFramework>net6.0</TargetFramework>
 	</PropertyGroup>
 </Project>
