@@ -9,6 +9,6 @@
 	</ItemGroup>
 	<PropertyGroup>
 		<!-- <TargetFramework>netcoreapp3.1</TargetFramework> -->
-		    <TargetFramework>net5.0</TargetFramework>
+		    <TargetFramework>net6.0</TargetFramework>
 	</PropertyGroup>
 </Project>
