@@ -1,73 +1,104 @@
 using System;
-using System.Linq;
 using System.Security.Cryptography;
-using System.Text;
-using Microsoft.Extensions.Options;
+using Microsoft.AspNetCore.Cryptography.KeyDerivation;
 
-namespace Servicio.Entidad.Helpers
+namespace Servicio.Entidad.Helpers;
+
+public sealed class PasswordHasher
 {
-    public sealed class PasswordHasher 
-    {
-        public string MD5(string word) 
-        {
-            MD5 md5 = MD5CryptoServiceProvider.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = md5.ComputeHash(encoding.GetBytes(word));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
-        }
+    public const int Pbkdf2Iterations = 1000;
 
-        public string Base64Encode(string word) 
+
+    public static string HashPasswordV3(string password)
+    {
+        return Convert.ToBase64String(HashPasswordV3(password, RandomNumberGenerator.Create()
+            , prf: KeyDerivationPrf.HMACSHA512, iterCount: Pbkdf2Iterations, saltSize: 128 / 8
+            , numBytesRequested: 256 / 8));
+    }
+
+
+    public static bool VerifyHashedPasswordV3(string hashedPasswordStr, string password)
+    {
+        byte[] hashedPassword = Convert.FromBase64String(hashedPasswordStr);
+        var iterCount = default(int);
+        var prf = default(KeyDerivationPrf);
+
+        try
         {
-            byte[] byt = System.Text.Encoding.UTF8.GetBytes(word);
-            return Convert.ToBase64String(byt);
+            // Read header information
+            prf = (KeyDerivationPrf)ReadNetworkByteOrder(hashedPassword, 1);
+            iterCount = (int)ReadNetworkByteOrder(hashedPassword, 5);
+            int saltLength = (int)ReadNetworkByteOrder(hashedPassword, 9);
+
+            // Read the salt: must be >= 128 bits
+            if (saltLength < 128 / 8)
+            {
+                return false;
+            }
+            byte[] salt = new byte[saltLength];
+            Buffer.BlockCopy(hashedPassword, 13, salt, 0, salt.Length);
+
+            // Read the subkey (the rest of the payload): must be >= 128 bits
+            int subkeyLength = hashedPassword.Length - 13 - salt.Length;
+            if (subkeyLength < 128 / 8)
+            {
+                return false;
+            }
+            byte[] expectedSubkey = new byte[subkeyLength];
+            Buffer.BlockCopy(hashedPassword, 13 + salt.Length, expectedSubkey, 0, expectedSubkey.Length);
+
+            // Hash the incoming password and verify it
+            byte[] actualSubkey = KeyDerivation.Pbkdf2(password, salt, prf, iterCount, subkeyLength);
+#if NETSTANDARD2_0 || NETFRAMEWORK
+            return ByteArraysEqual(actualSubkey, expectedSubkey);
+#elif NETCOREAPP
+            return CryptographicOperations.FixedTimeEquals(actualSubkey, expectedSubkey);
+#else
+#error Update target frameworks
+#endif
         }
-        public string Base64Decode(string word)
+        catch
         {
-            byte[] b = Convert.FromBase64String(word);
-            return System.Text.Encoding.UTF8.GetString(b);
-        }
-        public string SHA1(string str)
-        {
-            SHA1 sha1 = SHA1Managed.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = sha1.ComputeHash(encoding.GetBytes(str));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
-        }
-        public string SHA256(string str)
-        {
-            SHA256 sha256 = SHA256Managed.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = sha256.ComputeHash(encoding.GetBytes(str));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
-        }
-        public string SHA384(string str)
-        {
-            SHA384 sha384 = SHA384Managed.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = sha384.ComputeHash(encoding.GetBytes(str));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
-        }
-        public string SHA512(string str)
-        {
-            SHA512 sha512 = SHA512Managed.Create();
-            ASCIIEncoding encoding = new ASCIIEncoding();
-            byte[] stream = null;
-            StringBuilder sb = new StringBuilder();
-            stream = sha512.ComputeHash(encoding.GetBytes(str));
-            for (int i = 0; i < stream.Length; i++) sb.AppendFormat("{0:x2}", stream[i]);
-            return sb.ToString();
+            // This should never occur except in the case of a malformed payload, where
+            // we might go off the end of the array. Regardless, a malformed payload
+            // implies verification failed.
+            return false;
         }
     }
+
+
+    // privates
+    private static byte[] HashPasswordV3(string password, RandomNumberGenerator rng, KeyDerivationPrf prf, int iterCount, int saltSize, int numBytesRequested)
+    {
+        byte[] salt = new byte[saltSize];
+        rng.GetBytes(salt);
+        byte[] subkey = KeyDerivation.Pbkdf2(password, salt, prf, iterCount, numBytesRequested);
+        var outputBytes = new byte[13 + salt.Length + subkey.Length];
+        outputBytes[0] = 0x01; // format marker
+        WriteNetworkByteOrder(outputBytes, 1, (uint)prf);
+        WriteNetworkByteOrder(outputBytes, 5, (uint)iterCount);
+        WriteNetworkByteOrder(outputBytes, 9, (uint)saltSize);
+        Buffer.BlockCopy(salt, 0, outputBytes, 13, salt.Length);
+        Buffer.BlockCopy(subkey, 0, outputBytes, 13 + saltSize, subkey.Length);
+        return outputBytes;
+    }
+
+    private static void WriteNetworkByteOrder(byte[] buffer, int offset, uint value)
+    {
+        buffer[offset + 0] = (byte)(value >> 24);
+        buffer[offset + 1] = (byte)(value >> 16);
+        buffer[offset + 2] = (byte)(value >> 8);
+        buffer[offset + 3] = (byte)(value >> 0);
+    }
+
+    private static uint ReadNetworkByteOrder(byte[] buffer, int offset)
+    {
+        return ((uint)(buffer[offset + 0]) << 24)
+            | ((uint)(buffer[offset + 1]) << 16)
+            | ((uint)(buffer[offset + 2]) << 8)
+            | ((uint)(buffer[offset + 3]));
+    }
+
 }
+
+// https://stackoverflow.com/questions/20621950/asp-net-identitys-default-password-hasher-how-does-it-work-and-is-it-secure/72429730#72429730
