@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
+
 using Servicio.Entidad.Helpers;
 using Servicio.Entidad.interfaces;
 
@@ -8,12 +8,9 @@ namespace Servicio.Entidad.Models.Seguridad
 {
     public class Usuarios   : IEntity
     {
-        public Usuarios() => _passwordHasher = new PasswordHasher();
 
         #region Propiedades_privadas
 
-        // [NotMapped]
-        private PasswordHasher _passwordHasher{ get; set;}
 
         // [NotMapped]
         private string _xcontrasena { get; set; }
@@ -24,9 +21,10 @@ namespace Servicio.Entidad.Models.Seguridad
         public string xusuario { get; set; }
         public string nombres { get; set; }
         public string apellidos { get; set; }
-        public string xcontrasena { get => _passwordHasher.MD5(_xcontrasena); set =>  _xcontrasena = value; }
+        public string xcontrasena { get => PasswordHasher.HashPasswordV3(_xcontrasena); set =>  _xcontrasena = value; }
         public string correo { get; set; }
         public bool esActivo { get; set; }
         public IList<UsuariosRoles> UsuariosRoles { get; set; }
     }
 }
+
